Added Fade-Out for Menu
Added a fade-out for when a button is clicked. It stops further click detection, and returns the appropriate value after fade-out has completed.

It might be worth creating a fade-out class or something similar, so that these member variables do not need to be copied everywhere a fade-out is required.