Begun work on gameplay
A player has been built, and can only move within the game field. There is infrastructure in place for levels to affect the game state. A pause menu has been outlined briefly to be built upon later.

This pause screen will be a transparent overlay with two buttons, continue and return to menu.