Environment additions, sprite class
Implemented scrolling road through addition of a generic sprite class - this could be refined using the flyweight pattern to prevent loading the two road textures into memory when only one is required.

Some other issues are present, and some approaches are not efficient. These might not be totally addressed soon, as they are not critical.